Removed log entry from function header and updated reference.
Log entries to file/function headers should be avoided.  This is what version control logs are for.

The reference was updated for new constants added by Simon.  I put the reference info in the README with the others and added a tag based on the reference file name like the others.